Reuse _guess_format
